Compile playground with stubbed event impl
